Write a function that returns a key with the biggest integer value.

+++ 0x04-python-more_data_structures/10-best_score.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python3
 def best_score(a_dictionary):
     if a_dictionary:
-        max_key, max_value = "", int('-inf')
+        max_key, max_value = "", float('-inf')
         for key, value in dict_items:
             if value > max_value:
                 max_key, max_value = key, value
